uninsane
to my IRL friends at Anduril, especially the one who's been with me for
a decade: please understand i have no beef with you personally. i
respect your ability to navigate the many ambiguities and contradictions
of life. that any argument against Anduril as an entity will be an
argument you disagree with is testament to this: if this were otherwise,
then i know that you would not be where you are.

but when a rather meaningful assembly of NixOS contributors are
uncomfortable with military sponsorship/promotion, and you're made aware
of this, yet continue with it anyway, that's agitation. that's damaging
to this community.

by all means do seek to understand the diversity of perspectives within
this space and share your own. but you can and should recognize actions
as harmful even without understanding the entirety of why, just from the
consequentalist angle. i believe this letter makes that clear. i hope
that you can (if not now, then in time) be at home as an individual
of this community, while simply not promoting your employer within it.
much as i distance myself from my own government and overlook my
collaborator's government in order to be a part of this instead.